# Implement findPrevNextBeats to reduce Beats mutex locking.