--- a/src/track/beatgrid.h
+++ b/src/track/beatgrid.h
@@ -47,6 +47,7 @@ class BeatGrid : public QObject, public virtual Beats {
 
     virtual double findNextBeat(double dSamples) const;
     virtual double findPrevBeat(double dSamples) const;
+    virtual QPair<double, double> findPrevNextBeats(double dSamples) const;
     virtual double findClosestBeat(double dSamples) const;
     virtual double findNthBeat(double dSamples, int n) const;
     virtual BeatIterator* findBeats(double startSample, double stopSample) const;