--- a/src/track/beatgrid.h
+++ b/src/track/beatgrid.h
@@ -47,7 +47,9 @@ class BeatGrid : public QObject, public virtual Beats {
 
     virtual double findNextBeat(double dSamples) const;
     virtual double findPrevBeat(double dSamples) const;
-    virtual QPair<double, double> findPrevNextBeats(double dSamples) const;
+    virtual bool findPrevNextBeats(double dSamples,
+                                   double* dpPrevBeatSamples,
+                                   double* dpNextBeatSamples) const;
     virtual double findClosestBeat(double dSamples) const;
     virtual double findNthBeat(double dSamples, int n) const;
     virtual BeatIterator* findBeats(double startSample, double stopSample) const;