--- a/src/track/beatmap.cpp
+++ b/src/track/beatmap.cpp
@@ -22,7 +22,7 @@ inline double samplesToFrames(const double samples) {
 }
 
 inline double framesToSamples(const double frames) {
-    return frames * kFrameSize;
+    return floor(frames) * kFrameSize;
 }
 
 bool BeatLessThan(const Beat& beat1, const Beat& beat2) {
@@ -176,8 +176,12 @@ double BeatMap::findClosestBeat(double dSamples) const {
     if (!isValid()) {
         return -1;
     }
-    double nextBeat = findNextBeat(dSamples);
-    double prevBeat = findPrevBeat(dSamples);
+    QPair<double, double> beat_pair = findPrevNextBeats(dSamples);
+    if (beat_pair.first == -1 || beat_pair.second == -1) {
+        return -1;
+    }
+    double nextBeat = beat_pair.first;
+    double prevBeat = beat_pair.second;
     return (nextBeat - dSamples > dSamples - prevBeat) ? prevBeat : nextBeat;
 }
 
@@ -266,6 +270,89 @@ double BeatMap::findNthBeat(double dSamples, int n) const {
         }
     }
     return -1;
+}
+
+QPair<double, double> BeatMap::findPrevNextBeats(double dSamples) const {
+    QMutexLocker locker(&m_mutex);
+
+    if (!isValid()) {
+        return qMakePair(-1.0, -1.0);
+    }
+
+    Beat beat;
+    // Reduce sample offset to a frame offset.
+    beat.set_frame_position(samplesToFrames(dSamples));
+
+    // it points at the first occurence of beat or the next largest beat
+    BeatList::const_iterator it =
+            qLowerBound(m_beats.begin(), m_beats.end(), beat, BeatLessThan);
+
+    // If the position is within 1/10th of a second of the next or previous
+    // beat, pretend we are on that beat.
+    const double kFrameEpsilon = 0.1 * m_iSampleRate;
+
+    // Back-up by one.
+    if (it != m_beats.begin()) {
+        --it;
+    }
+
+    // Scan forward to find whether we are on a beat.
+    BeatList::const_iterator on_beat = m_beats.end();
+    BeatList::const_iterator previous_beat = m_beats.end();
+    BeatList::const_iterator next_beat = m_beats.end();
+    for (; it != m_beats.end(); ++it) {
+        qint32 delta = it->frame_position() - beat.frame_position();
+
+        // We are "on" this beat.
+        if (abs(delta) < kFrameEpsilon) {
+            on_beat = it;
+            break;
+        }
+
+        if (delta < 0) {
+            // If we are not on the beat and delta < 0 then this beat comes
+            // before our current position.
+            previous_beat = it;
+        } else {
+            // If we are past the beat and we aren't on it then this beat comes
+            // after our current position.
+            next_beat = it;
+            // Stop because we have everything we need now.
+            break;
+        }
+    }
+
+    // If we are within epsilon samples of a beat then the immediately next and
+    // previous beats are the beat we are on.
+    if (on_beat != m_beats.end()) {
+        previous_beat = on_beat;
+        next_beat = on_beat + 1;
+    }
+
+    double dPrevSamples = -1;
+    double dNextSamples = -1;
+
+    for (; next_beat != m_beats.end(); ++next_beat) {
+        if (!next_beat->enabled()) {
+            continue;
+        }
+        dNextSamples = framesToSamples(next_beat->frame_position());
+        break;
+    }
+    if (previous_beat != m_beats.end()) {
+        for (; true; --previous_beat) {
+            if (previous_beat->enabled()) {
+                dPrevSamples = framesToSamples(previous_beat->frame_position());
+                break;
+            }
+
+            // Don't step before the start of the list.
+            if (previous_beat == m_beats.begin()) {
+                break;
+            }
+        }
+    }
+    return qMakePair(dPrevSamples, dNextSamples);
 }
 
 BeatIterator* BeatMap::findBeats(double startSample, double stopSample) const {