--- a/src/track/beatmap.h
+++ b/src/track/beatmap.h
@@ -54,6 +54,7 @@ class BeatMap : public QObject, public Beats {
 
     virtual double findNextBeat(double dSamples) const;
     virtual double findPrevBeat(double dSamples) const;
+    virtual QPair<double, double> findPrevNextBeats(double dSamples) const;
     virtual double findClosestBeat(double dSamples) const;
     virtual double findNthBeat(double dSamples, int n) const;
     virtual BeatIterator* findBeats(double startSample, double stopSample) const;