--- a/src/track/beats.h
+++ b/src/track/beats.h
@@ -61,8 +61,16 @@ class Beats {
     // beat, dSamples is returned.
     virtual double findPrevBeat(double dSamples) const = 0;
 
+    // Starting from sample dSamples, return the sample of the previous beat
+    // and next beat.  Either can be -1 if none exists.  If dSamples refers
+    // to the location of the beat, the first value is dSamples, and the second
+    // value is the next beat position.  Non- -1 values are guaranteed to be
+    // even.
+    virtual QPair<double, double> findPrevNextBeats(double dSamples) const = 0;
+
     // Starting from sample dSamples, return the sample of the closest beat in
-    // the track, or -1 if none exists.
+    // the track, or -1 if none exists.  Non- -1 values are guaranteed to be
+    // even.
     virtual double findClosestBeat(double dSamples) const = 0;
 
     // Find the Nth beat from sample dSamples. Works with both positive and