… src/test/beatgridtest.cpp
… src/test/beatmaptest.cpp