Change paired calls to findnext and findprev to use the new function.
Added tests to confirm it works as intended.