@ywwg 2015-02-19
Interestingly, as a consequence, as a result there are no actual calls to findPrevBeat in the code (other than tests)