@ywwg 2015-02-20
with this, and then a subsequent change that reduces calls to findnthbeat even further, I get the following change in --developer playing two sync locked tracks, one beatgrid, one beatmap, for about a minute:

before:
Debug [Main]: Stat("EngineMaster::processChannels","count=3438,sum=7.60152e+08ns,average=221103ns,min=4521ns,max=998917ns,variance=2.18565e+10ns^2,stddev=147839ns")
after:
Debug [Main]: Stat("EngineMaster::processChannels","count=3391,sum=5.61576e+08ns,average=165608ns,min=4711ns,max=919647ns,variance=1.17827e+10ns^2,stddev=108548ns")

which is a 25% decrease in average time.  Also much better worst-case performance and better standard deviation (though still bad -- close to 1ms).