@rryan 2015-02-22
lgtm other than minor comments

nice!