@ywwg 2015-02-22
notes addressed -- merging and then pushing the second half of this change